OnHit & OnTakeDamage changes
Pray

+++ Assets/Scripts/StatusEffects/DamageOverTime/DamageOverTimeStatus.cs
@@ -23,7 +23,7 @@ public class DamageOverTimeStatus : StatusEffect
         // Temporary solution for between refactoring
         else
         {
-            handler.GetComponentInParent<NetworkPlayer_Health>().OnTakeDamage((int)damage);
+            handler.GetComponentInParent<NetworkPlayer_Health>().OnTakeDamage(damage);
         }
     }
